+++ src/main/java/com/example/yblog/portfolio/service/PortfolioBoardService.java
@@ -15,6 +15,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -51,8 +52,7 @@ public class PortfolioBoardService {
             portfolioRepository.save(portfolioBoard);
 
 
-
-
+            deletetemporaryStorage(principal.getUsername(), true);
 
         }else{
             System.out.println("do not Save because no title");
@@ -75,15 +75,27 @@ public class PortfolioBoardService {
                 .orElseThrow(()->{
                     return new IllegalArgumentException("포트폴리오 게시판 수정실패 아이디를 찾을수 없습니다.");
                 });
+
+        // 글과 사진 수정, 추가
+        PortfolioBoard portfolioBoard1 = ImageSearch(eternalPortfolioBoard.getUser().getUsername(),
+                portfolioBoard,eternalPortfolioBoard,true);
+
         eternalPortfolioBoard.setTitle(portfolioBoard.getTitle());
-        eternalPortfolioBoard.setContent(portfolioBoard.getContent());
+        eternalPortfolioBoard.setContent(portfolioBoard1.getContent());
         // 더티체킹
 
     }
 
     @Transactional
     public  void portBoardDelete(int portboardId){
-        portfolioRepository.deleteById(portboardId);
+        Optional<PortfolioBoard> portfolioBoard = portfolioRepository.findById(portboardId);
+
+        if (portfolioBoard.isEmpty()){
+            System.out.println("portboard id null");
+        }else{
+            deleteConfirmfile(portfolioBoard.get().getImagefileid());
+            portfolioRepository.deleteById(portboardId);
+        }
     }
 
 
@@ -93,7 +105,7 @@ public class PortfolioBoardService {
     // 2. contents 안에있는 사진 src값을 변경 (board를)
     // 3. 임시파일 사진들을 전부 삭제
     // modify => 수정이면 참, 거짓이면 새로생성
-    public PortfolioBoard ImageSearch(String username , PortfolioBoard yBoard, PortfolioBoard yBoard1, boolean modify)  {
+    public PortfolioBoard ImageSearch(String username , PortfolioBoard portfolioBoard, PortfolioBoard yBoard1, boolean modify)  {
         String temporary = "C:\\temporary_storage\\"+username; // 임시 파일 저장경로
         String fileRoot =null;
         String BoardTitleUUID =null;
@@ -106,7 +118,7 @@ public class PortfolioBoardService {
             chfileRoot = "/Confirm_SaveImage/"+BoardTitleUUID+"/";
         }else{
             // 지금 새로만드는 파일이니깐 새로운 랜덤 키값을 지정
-            BoardTitleUUID = UUID.randomUUID()+"-"+yBoard.getUser().getId();
+            BoardTitleUUID = UUID.randomUUID()+"-"+portfolioBoard.getUser().getId();
             fileRoot = "C:\\Confirm_SaveImage\\"+BoardTitleUUID+"\\";	//저장확정 경로
             chfileRoot = "/Confirm_SaveImage/"+BoardTitleUUID+"/";
         }
@@ -144,19 +156,73 @@ public class PortfolioBoardService {
         }
 
 
-        System.out.println(yBoard.getContent());
-        String ChContent = yBoard.getContent().replace("/temporary_storage/"+username+"/", chfileRoot);
+        System.out.println(portfolioBoard.getContent());
+        String ChContent = portfolioBoard.getContent().replace("/temporary_storage/"+username+"/", chfileRoot);
         System.out.println(ChContent);
 
-        yBoard.setContent(ChContent);
-        yBoard.setImagefileid(BoardTitleUUID);
+        portfolioBoard.setContent(ChContent);
+        portfolioBoard.setImagefileid(BoardTitleUUID);
 
-        return yBoard;
+        return portfolioBoard;
+    }
+
+
+    public void deletetemporaryStorage(String username, boolean writebool){
+        String path = "C:\\temporary_storage\\"+username;
+        File folder = new File(path);
+        try {
+            while(folder.exists()) {
+                File[] folder_list = folder.listFiles(); //파일리스트 얻어오기
+
+                for (int j = 0; j < folder_list.length; j++) {
+                    folder_list[j].delete(); //파일 삭제
+                    System.out.println("All delete Temporary File UserName : "+username);
+
+                }
+
+                if(folder_list.length == 0 && folder.isDirectory()){
+                    folder.delete(); //대상폴더 삭제
+                    System.out.println("All delete Temporary Folder UserName : "+username);
+                }
+            }
+        } catch (Exception e) {
+            //참이면 글쓴거임
+            if(writebool){
+                System.out.println("write after delete imageFile");
+            }else{
+                // else면 글을 쓰지않고 나갓을때
+                e.getStackTrace();
+            }
+
+
+        }
     }
 
 
 
+    public void deleteConfirmfile(String filename){
+        String path = "C:\\Confirm_SaveImage\\"+filename;
+        File folder = new File(path);
+        try {
+            while(folder.exists()) {
+                File[] folder_list = folder.listFiles(); //파일리스트 얻어오기
 
+                for (int j = 0; j < folder_list.length; j++) {
+                    folder_list[j].delete(); //파일 삭제
+                    System.out.println("All delete Temporary File UserName : "+filename);
+
+                }
+
+                if(folder_list.length == 0 && folder.isDirectory()){
+                    folder.delete(); //대상폴더 삭제
+                    System.out.println("All delete Temporary Folder UserName : "+filename);
+                }
+            }
+        } catch (Exception e) {
+            e.getStackTrace();
+
+        }
+    }
